+++ src/plugin.js
@@ -23,13 +23,24 @@
 
 const fp = require('fastify-plugin')
 const { CloudEvent, CloudEventTransformer, JSONBatch } = require('cloudevent') // get CloudEvent definition and related utilities
+const {
+  ensureIsString,
+  ensureIsBoolean,
+  ensureIsObject,
+  ensureIsObjectPlain,
+  ensureIsFunction,
+  idMaker,
+  isDefinedAndNotNull,
+  isValue
+} = require('./utils')
+const setupHooks = require('./plugin-hooks')
 
 const pluginName = require('../package.json').name // get plugin name
 const pluginVersion = require('../package.json').version // get plugin version
 
 /**
  * Plugin implementation.
- * Note that's an async function.
+ * Note that it's an async function.
  *
  * @param {!object} fastify Fastify instance
  * @param {object} [options={}] plugin configuration options
@@ -38,13 +49,15 @@ const pluginVersion = require('../package.json').version // get plugin version
  *     <li>cloudEventExtensions (object, default null) Extensions for events generated,</li>
  *     <li>cloudEventOptions (object, default empty) Options for events generated,</li>
  *     <li>idGenerator (function, default idMaker) to build ID for events generated,</li>
- *     <li>includeHeaders (boolean, default false) flag to enable the add HTTP request Headers in events generated,</li>
+ *     <li>includeHeaders (boolean, default false) flag to enable adding HTTP request Headers in events generated,</li>
  *     <li>includeHttpAttributes (boolean, default false) flag to add some HTTP request attributes in events generated,</li>
- *     <li>includeRedundantAttributes` (boolean, default false) flag to add some redundant attributes inside data in events generated,</li>
+ *     <li>includeRedundantAttributes (boolean, default false) flag to add some redundant attributes inside data in events generated,</li>
  *     <li>onCloseCallback (function, no default) callback function for the 'onClose' hook,</li>
  *     <li>onErrorCallback (function, no default) callback function for the 'onError' hook,</li>
+ *     <li>onListenCallback (function, no default) callback function for the 'onListen' hook,</li>
  *     <li>onReadyCallback (function, no default) callback function for the 'onReady' hook,</li>
  *     <li>onRegisterCallback (function, no default) callback function for the 'onRegister' hook,</li>
+ *     <li>onRequestAbortCallback (function, no default) callback function for the 'onRequestAbort' hook,</li>
  *     <li>onRequestCallback (function, no default) callback function for the 'onRequest' hook,</li>
  *     <li>onResponseCallback (function, no default) callback function for the 'onResponse' hook,</li>
  *     <li>onRouteCallback (function, no default) callback function for the 'onRoute' hook,</li>
@@ -73,8 +86,10 @@ async function fastifyCloudEvents (fastify, options) {
     includeRedundantAttributes = false,
     onCloseCallback = null,
     onErrorCallback = null,
+    onListenCallback = null,
     onReadyCallback = null,
     onRegisterCallback = null,
+    onRequestAbortCallback = null,
     onRequestCallback = null,
     onResponseCallback = null,
     onRouteCallback = null,
@@ -98,19 +113,21 @@ async function fastifyCloudEvents (fastify, options) {
   ensureIsBoolean(includeHeaders, 'includeHeaders')
   ensureIsBoolean(includeHttpAttributes, 'includeHttpAttributes')
   ensureIsBoolean(includeRedundantAttributes, 'includeRedundantAttributes')
-  ensureIsFunction(onRequestCallback, 'onRequestCallback')
-  ensureIsFunction(preParsingCallback, 'preParsingCallback')
-  ensureIsFunction(preValidationCallback, 'preValidationCallback')
-  ensureIsFunction(preHandlerCallback, 'preHandlerCallback')
-  ensureIsFunction(preSerializationCallback, 'preSerializationCallback')
-  ensureIsFunction(onErrorCallback, 'onErrorCallback')
-  ensureIsFunction(onSendCallback, 'onSendCallback')
-  ensureIsFunction(onResponseCallback, 'onResponseCallback')
   ensureIsFunction(onCloseCallback, 'onCloseCallback')
-  ensureIsFunction(onRouteCallback, 'onRouteCallback')
-  ensureIsFunction(onRegisterCallback, 'onRegisterCallback')
+  ensureIsFunction(onErrorCallback, 'onErrorCallback')
+  ensureIsFunction(onListenCallback, 'onListenCallback')
   ensureIsFunction(onReadyCallback, 'onReadyCallback')
+  ensureIsFunction(onRegisterCallback, 'onRegisterCallback')
+  ensureIsFunction(onRequestAbortCallback, 'onRequestAbortCallback')
+  ensureIsFunction(onRequestCallback, 'onRequestCallback')
+  ensureIsFunction(onResponseCallback, 'onResponseCallback')
+  ensureIsFunction(onRouteCallback, 'onRouteCallback')
+  ensureIsFunction(onSendCallback, 'onSendCallback')
   ensureIsFunction(onTimeoutCallback, 'onTimeoutCallback')
+  ensureIsFunction(preHandlerCallback, 'preHandlerCallback')
+  ensureIsFunction(preParsingCallback, 'preParsingCallback')
+  ensureIsFunction(preSerializationCallback, 'preSerializationCallback')
+  ensureIsFunction(preValidationCallback, 'preValidationCallback')
   ensureIsObjectPlain(cloudEventOptions, 'cloudEventOptions')
   ensureIsObjectPlain(cloudEventExtensions, 'cloudEventExtensions')
 
@@ -271,213 +288,10 @@ async function fastifyCloudEvents (fastify, options) {
     cloudEventExtensions
   })
 
-  // handle hooks, only when related callback are defined
-  // see [Hooks - Fastify reference - GitHub](https://github.com/fastify/fastify/blob/main/docs/Reference/Hooks.md)
-
-  if (onRequestCallback !== null) {
-    fastify.addHook('onRequest', async (request, reply) => {
-      // small optimization: pass a null reply because no useful here
-      const ce = builders.buildCloudEventForHook('onRequest', request, null)
-      // add some http related attributes to data, could be useful to have
-      if (includeHttpAttributes !== null && includeHttpAttributes === true) {
-        ce.data.request.httpVersion = request.raw.httpVersion
-        ce.data.request.originalUrl = request.raw.originalUrl
-        ce.data.request.upgrade = request.raw.upgrade
-      }
-      // console.log(`DEBUG - onRequest: created CloudEvent ${CloudEventTransformer.dumpObject(ce, 'ce')}`)
-      // send the event to the callback
-      await onRequestCallback(ce)
-    })
-  }
-
-  if (preParsingCallback !== null) {
-    fastify.addHook('preParsing', async (request, reply, payload) => {
-      const ce = builders.buildCloudEventForHook('preParsing', request, reply, // payload)
-        null) // do not pass payload here or a "Converting circular structure to JSON" will be raised if enabled ...
-      await preParsingCallback(ce)
-    })
-  }
-
-  if (preValidationCallback !== null) {
-    fastify.addHook('preValidation', async (request, reply) => {
-      const ce = builders.buildCloudEventForHook('preValidation', request, reply)
-      await preValidationCallback(ce)
-    })
-  }
-
-  if (preHandlerCallback !== null) {
-    fastify.addHook('preHandler', async (request, reply) => {
-      const ce = builders.buildCloudEventForHook('preHandler', request, reply)
-      await preHandlerCallback(ce)
-    })
-  }
-
-  if (preSerializationCallback !== null) {
-    fastify.addHook('preSerialization', async (request, reply, payload) => {
-      const ce = builders.buildCloudEventForHook('preSerialization', request, reply, payload)
-      await preSerializationCallback(ce)
-    })
-  }
-
-  if (onErrorCallback !== null) {
-    fastify.addHook('onError', async (request, reply, error) => {
-      const processInfoAsData = CloudEventTransformer.processInfoToData()
-      const errorAsData = CloudEventTransformer.errorToData(error, {
-        includeStackTrace: true,
-        addStatus: true,
-        addTimestamp: true
-      })
-      const ceData = {
-        request: builders.buildRequestDataForCE(request),
-        reply: builders.buildReplyDataForCE(reply),
-        error: errorAsData,
-        process: processInfoAsData
-      }
-      if (includeRedundantAttributes !== null && includeRedundantAttributes === true) {
-        ceData.id = request.id
-        ceData.timestamp = CloudEventTransformer.timestampToNumber()
-      }
-      const ce = new fastify.CloudEvent(idGenerator.next().value,
-        `${baseNamespace}.onError`,
-        builders.buildSourceUrl(request.url),
-        ceData,
-        cloudEventOptions,
-        cloudEventExtensions
-      )
-      await onErrorCallback(ce)
-      // done() // do not pass the error to the done callback here
-    })
-  }
-
-  if (onSendCallback !== null) {
-    fastify.addHook('onSend', async (request, reply, payload) => {
-      const ce = builders.buildCloudEventForHook('onSend', request, reply, payload)
-      await onSendCallback(ce)
-    })
-  }
-
-  if (onResponseCallback !== null) {
-    fastify.addHook('onResponse', async (request, reply) => {
-      const ce = builders.buildCloudEventForHook('onResponse', request, reply)
-      // keep the request attribute from data, even if more data will be shown here
-      await onResponseCallback(ce)
-    })
-  }
-
-  if (onTimeoutCallback !== null) {
-    fastify.addHook('onTimeout', async (request, reply) => {
-      const ce = builders.buildCloudEventForHook('onTimeout', request, reply)
-      // keep the request attribute from data, even if more data will be shown here
-      await onTimeoutCallback(ce)
-    })
-  }
-
-  if (onCloseCallback !== null) {
-    // hook to plugin shutdown
-    fastify.addHook('onClose', async (instance) => {
-      const ce = new fastify.CloudEvent(idGenerator.next().value,
-        `${baseNamespace}.onClose`,
-        builders.buildSourceUrl(),
-        builders.buildPluginDataForCE('plugin shutdown'), // data
-        cloudEventOptions,
-        cloudEventExtensions
-      )
-      await onCloseCallback(ce)
-    })
-  }
-
-  if (onRouteCallback !== null) {
-    fastify.addHook('onRoute', (routeOptions) => {
-      const ce = new fastify.CloudEvent(idGenerator.next().value,
-        `${baseNamespace}.onRoute`,
-        builders.buildSourceUrl(),
-        routeOptions, // data
-        cloudEventOptions,
-        cloudEventExtensions
-      )
-      onRouteCallback(ce)
-    })
-  }
-
-  if (onRegisterCallback !== null) {
-    fastify.addHook('onRegister', (instance, opts) => {
-      const ce = new fastify.CloudEvent(idGenerator.next().value,
-        `${baseNamespace}.onRegister`,
-        builders.buildSourceUrl(),
-        builders.buildPluginDataForCE(`plugin registration, with options: ${JSON.stringify(opts)}`), // data
-        cloudEventOptions,
-        cloudEventExtensions
-      )
-      onRegisterCallback(ce)
-    })
-  }
-
-  if (onReadyCallback !== null) {
-    // triggered before the server starts listening for requests
-    fastify.addHook('onReady', async () => {
-      const ce = new fastify.CloudEvent(idGenerator.next().value,
-        `${baseNamespace}.onReady`,
-        builders.buildSourceUrl(),
-        builders.buildPluginDataForCE('plugin ready'), // data
-        cloudEventOptions,
-        cloudEventExtensions
-      )
-      await onReadyCallback(ce)
-    })
-  }
+  // setup all hooks
+  setupHooks(fastify, options, builders)
 
   // done()
-}
-
-// utility functions
-
-function ensureIsString (arg, name = 'arg') {
-  if (arg !== null && typeof arg !== 'string') {
-    throw new TypeError(`The argument '${name}' must be a string, instead got a '${typeof arg}'`)
-  }
-}
-
-function ensureIsBoolean (arg, name = 'arg') {
-  if (arg !== null && typeof arg !== 'boolean') {
-    throw new TypeError(`The argument '${name}' must be a boolean, instead got a '${typeof arg}'`)
-  }
-}
-
-function ensureIsObject (arg, name = 'arg') {
-  if (arg !== null && typeof arg !== 'object') {
-    throw new TypeError(`The argument '${name}' must be a object, instead got a '${typeof arg}'`)
-  }
-}
-
-function ensureIsObjectPlain (arg, name = 'arg') {
-  if (arg !== null && Object.prototype.toString.call(arg) === '[object Object]') {
-    return new TypeError(`The argument '${name}' must be a plain object, instead got a '${typeof arg}'`)
-  }
-}
-
-function ensureIsFunction (arg, name = 'arg') {
-  if (arg !== null && typeof arg !== 'function') {
-    throw new TypeError(`The argument '${name}' must be a function, instead got a '${typeof arg}'`)
-  }
-}
-
-const hostname = require('node:os').hostname()
-const idPrefix = `fastify@${hostname}`
-function * idMaker () {
-  while (true) {
-    const timestamp = CloudEventTransformer.timestampToNumber()
-    yield `${idPrefix}@${timestamp}`
-  }
-}
-
-function isDefinedAndNotNull (arg) {
-  return (arg !== undefined && arg !== null)
-}
-
-function isValue (arg) {
-  return ((arg !== undefined && arg !== null) &&
-    (typeof arg === 'string' || typeof arg === 'boolean' || (typeof arg === 'number' && !isNaN(arg)))
-  )
 }
 
 module.exports = fp(fastifyCloudEvents, {